add correct key to .env and .env example
