+++ .envexample
@@ -1,3 +1,3 @@
 PORT=4000
 mapAPI=<secretKey>
-api=http://localhost:8080/
+DEV_API=http://localhost:8080/
