+++ config/index.js
@@ -1,9 +1,6 @@
 module.exports = {
-    mapAPI: process.env.mapAPI,
-    api:
-        process.env.NODE_ENV === "development"
-            ? process.env.DEV_API
-            : process.env.PROD_API,
+	mapAPI: process.env.mapAPI,
+	api: process.env.NODE_ENV === 'development' ? process.env.DEV_API : process.env.PROD_API,
 
-    port: process.env.PORT,
+	port: process.env.PORT
 };
