+++ package.json
@@ -4,7 +4,9 @@
 	"description": "",
 	"main": "index.js",
 	"scripts": {
-		"start": "nodemon -r dotenv/config index.js"
+		"start": "per-env",
+		"start:development": "nodemon -r dotenv/config index.js",
+		"start:production": "node index.js"
 	},
 	"keywords": [],
 	"author": "",
@@ -14,6 +16,7 @@
 		"express": "^4.17.1",
 		"localStorage": "^1.0.4",
 		"node-fetch": "^2.6.0",
+		"per-env": "^1.0.2",
 		"pug": "^2.0.4"
 	},
 	"devDependencies": {
