+++ public/js/index.js
@@ -1,11 +1,12 @@
-const api = document.querySelector('link[rel="api]').href;
+import { api } from './utils.js';
 
 document.addEventListener('DOMContentLoaded', async () => {
 	// await the fetch call from the api backend
-	try {
-		const res = await fetch(`${api}`);
-		return await res.json();
-	} catch (e) {
-		console.error(e);
-	}
+	console.log('where is my api? ', api);
+	// try {
+	// 	const res = await fetch(`${api}`);
+	// 	return await res.json();
+	// } catch (e) {
+	// 	console.error(e);
+	// }
 });
