+++ public/js/log-in.js
@@ -1,4 +1,4 @@
-const api = document.querySelector('link[rel="api]').href;
+import { handleErrors, api } from './utils.js';
 
 if (localStorage.getItem('VIDEO_EATS_ACCESS_TOKEN')) {
 	alert('You are already logged in!');
@@ -35,26 +35,6 @@ logInForm.addEventListener('submit', async (e) => {
 		// redirect to home page to see all tweets:
 		window.location.href = '/';
 	} catch (err) {
-		if (err.status >= 400 && err.status < 600) {
-			const errorJSON = await err.json();
-			const errorsContainer = document.querySelector('.errors-container');
-			let errorsHtml = [
-				`<div class="alert alert-danger">
-                    Something went wrong. Please try again.
-                 </div>`
-			];
-			const { errors } = errorJSON;
-			if (errors && Array.isArray(errors)) {
-				errorsHtml = errors.map(
-					(message) =>
-						`<div class="alert alert-danger">
-                        ${message}
-                    </div>`
-				);
-			}
-			errorsContainer.innerHTML = errorsHtml.join('');
-		} else {
-			alert('Something went wrong. Please check your internet connection and try again!');
-		}
+		handleErrors(err);
 	}
 });
