+++ public/js/write-a-review.js
@@ -1,4 +1,4 @@
-const api = document.querySelector('link[rel="api]').href;
+import { handleErrors, api } from './utils.js';
 
 document.addEventListener('DOMContentLoaded', async () => {
 	const reviewForm = document.querySelector('.create-review');
@@ -101,30 +101,7 @@ document.addEventListener('DOMContentLoaded', async () => {
 			if (err.status === 401) {
 				window.location.href = '/log-in';
 			}
-			if (err.status >= 400 && err.status < 600) {
-				const errorJSON = await err.json();
-				const errorsContainer = document.querySelector('.errors-container');
-				let errorsHtml = [
-					`
-                    <div class="alert alert-danger">
-                        Something went wrong. Please try again.
-                    </div>
-                `
-				];
-				const { errors } = errorJSON;
-				if (errors && Array.isArray(errors)) {
-					errorsHtml = errors.map(
-						(message) => `
-                        <div class="alert alert-danger">
-                            ${message.message}
-                        </div>
-                        `
-					);
-				}
-				errorsContainer.innerHTML = errorsHtml.join('');
-			} else {
-				alert('Something went wrong. Please check your internet connection and try again!');
-			}
+			handleErrors(err);
 		}
 	});
