+++ views/business.pug
@@ -20,4 +20,4 @@ block content
                 p #{business.address}
             button.btn.btn-secondary(type="button" id="write-a-review-button") Write a Review
         .review-section
-    script(src="/js/business.js")
+    script(type="module" src="/js/business.js")
