+++ views/google-map.pug
@@ -25,7 +25,7 @@ doctype html
   </head>
   <body>
     <div id="map"></div>
-    script(src='/js/initMap.js')
+    script(type="module" src='/js/initMap.js')
     script(src=`https://maps.googleapis.com/maps/api/js?key=${mapAPI}&callback=initMap` async='' defer='')
   </body>
 </html>
