+++ views/index.pug
@@ -2,3 +2,4 @@ extends layout.pug
 
 block content
   h1 Hello
+  script(type="module" src="/js/index.js")
