+++ views/layout.pug
@@ -38,6 +38,6 @@ html
           //- TODO: alter search route to be something like /search/?= once we implement /search
     .body-content
       block content
-    script(src="/js/navbar-search.js")
-    script(src="/js/user-auth.js")
-    script(src="/js/log-out.js")
+    script(type="module" src="/js/navbar-search.js")
+    script(type="module" src="/js/user-auth.js")
+    script(type="module" src="/js/log-out.js")
