+++ views/log-in.pug
@@ -15,5 +15,5 @@ block content
       a(href='/sign-up') Don't have an account? Sign up here.
     div(class="image-container")
       img(src="/images/sign-up-image.jpg")
-  script(src="/js/log-in.js")
+  script(type="module" src="/js/log-in.js")
   //- TODO: Create a demo user login info
