+++ views/search.pug
@@ -4,4 +4,4 @@ block content
   //-includes google-map.pug
   h1 Search Results
   .business-card-container
-  script(src="/js/navbar-search.js")
+  script(type="module" src="/js/navbar-search.js")
