+++ views/sign-up.pug
@@ -21,4 +21,4 @@ block content
       a(href='/log-in' class="mt-2") Already have an account? Log in here.
     div(class="image-container")
       img(src="/images/sign-up-image.jpg")
-  script(src="/js/sign-up.js")
+  script(type="module" src="/js/sign-up.js")
