+++ views/write-a-review.pug
@@ -27,5 +27,5 @@ block content
                 button(type="submit" class="btn btn-primary") Post Review
                 button(type="button" class="ml-2 btn" id="reivew-cancel-button") Cancel
 
-    script(src="/js/write-a-review.js")
+    script(type="module" src="/js/write-a-review.js")
     //- script(src="/js/search-tags.js")
